moving all functionality of old 'helpers' components into new prices section
