+++ support-frontend/assets/pages/paper-subscription-landing/components/content/deliveryTab.jsx
@@ -6,9 +6,7 @@ import React from 'react';
 import FlexContainer from 'components/containers/flexContainer';
 import GridImage from 'components/gridImage/gridImage';
 import { setTab } from '../../paperSubscriptionLandingPageActions';
-import {
-  LinkTo,
-} from './helpers';
+import LinkTo from './linkTo';
 import { Collection } from 'helpers/productPrice/fulfilmentOptions';
 // import { paperHasDeliveryEnabled } from 'helpers/subscriptions';
 import { Accordion, AccordionRow } from '@guardian/src-accordion';
