+++ support-frontend/assets/pages/paper-subscription-landing/components/content/prices.jsx
@@ -6,14 +6,23 @@ import { space } from '@guardian/src-foundations';
 import { headline } from '@guardian/src-foundations/typography';
 import { from } from '@guardian/src-foundations/mq';
 import { SvgInfo } from '@guardian/src-icons';
+
 import ProductInfoChip from 'components/product/productInfoChip';
 import FlexContainer from 'components/containers/flexContainer';
 import ProductOption, { type Product } from 'components/product/productOption';
+import { Collection, HomeDelivery } from 'helpers/productPrice/fulfilmentOptions';
+import { Callout } from 'components/text/text';
+import { flashSaleIsActive, getDiscount, getDuration } from 'helpers/flashSale';
+
+import LinkTo from './linkTo';
+import { setTab } from '../../paperSubscriptionLandingPageActions';
 import { type ActiveTabState } from '../../paperSubscriptionLandingPageReducer';
 
 export type PropTypes = {|
   activeTab: ActiveTabState,
+  setTabAction: typeof setTab,
   products: Product[],
+  useDigitalVoucher: boolean,
 |};
 
 const pricesSection = css`
@@ -65,11 +74,28 @@ const pricesInfo = css`
   margin-top: ${space[6]}px;
 `;
 
-function Prices({ activeTab, products }: PropTypes) {
-  const infoText = `${activeTab === 'HomeDelivery' ? 'Delivery is included. ' : ''}You can cancel your subscription at any time`;
+const DiscountCalloutMaybe = () => {
+  if (!flashSaleIsActive('Paper', 'GBPCountries')) { return null; }
+  const [discount, duration] = [
+    getDiscount('Paper', 'GBPCountries'),
+    getDuration('Paper', 'GBPCountries'),
+  ];
+  if (discount && duration) {
+    return <Callout>Save up to {Math.round(discount * 100)}% for {duration}</Callout>;
+  } else if (discount) {
+    return <Callout>Save up to {Math.round(discount * 100)}% </Callout>;
+  }
+  return null;
+};
+
+function Prices({
+  activeTab, setTabAction, products, useDigitalVoucher,
+}: PropTypes) {
+  const infoText = `${activeTab === HomeDelivery ? 'Delivery is included. ' : ''}You can cancel your subscription at any time`;
   return (
     <section css={pricesSection} id="subscribe">
       <h2 css={pricesHeadline}>Pick your subscription package below</h2>
+      <DiscountCalloutMaybe />
       <FlexContainer cssOverrides={priceBoxes}>
         {products.map(product => (
           <ProductOption
@@ -85,6 +111,15 @@ function Prices({ activeTab, products }: PropTypes) {
           />
         ))}
       </FlexContainer>
+      <div css={pricesInfo}>
+        {
+            activeTab === Collection ?
+              <LinkTo tab={HomeDelivery} setTabAction={setTabAction}>Switch to Delivery</LinkTo> :
+              <LinkTo tab={Collection} setTabAction={setTabAction}>
+                  Switch to {useDigitalVoucher ? 'Subscription card' : 'Vouchers'}
+              </LinkTo>
+          }
+      </div>
       <div css={pricesInfo}>
         <ProductInfoChip icon={<SvgInfo />}>
           {infoText}
