+++ support-frontend/assets/pages/paper-subscription-landing/paperSubscriptionLandingPage.jsx
@@ -60,7 +60,6 @@ const content = (
         <CentredContainer>
           <Block>
             <Tabs />
-            {/* <TabsContent /> */}
           </Block>
         </CentredContainer>
       </FullWidthContainer>
