+++ support-frontend/stories/_index.js
@@ -11,10 +11,8 @@ module.exports = () => {
   require('./menu.jsx');
   require('./type.jsx');
   require('./layout.jsx');
-  require('./tabs.jsx');
   require('./animations.jsx');
   require('./contributionTicker.jsx');
-  require('./cards.jsx');
   require('./containers.jsx');
   require('./page.jsx');
   require('./product.jsx');
